% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/across.R
\name{across}
\alias{across}
\title{Apply a function to a set of columns (should be plugged in a select or an update)}
\usage{
across(f, ..., bylist = NULL, pref = "")
}
\arguments{
\item{f}{function to apply}

\item{...}{column names}

\item{bylist}{column names, if in a list and characters}

\item{pref}{string prefix to apply to the newly created column names}
}
\value{
an alist containing the recipe to construct the new columns
}
\description{
Apply a function to a set of columns (should be plugged in a select or an update)
}
\examples{
d <- data.frame(
date=seq(as.Date("2001-01-01"), as.Date("2001-10-01"), by="month"),
ticker=sample(c("1 HK", "2 HK"), 10, TRUE),
ret1=rnorm(10),
ret2=rnorm(10),
ret3=rnorm(10)
)

d \%>\% select(across(mean, ret1, ret2, pref="av_"), by = ticker)

mycols <- c("ret1", "ret2")
d \%>\% select(across(mean, bylist = mycols, pref="av_"), by = ticker)
}
