% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aj.R
\name{aj}
\alias{aj}
\title{As of join}
\usage{
aj(..., x, y)
}
\arguments{
\item{x}{data.frame}

\item{y}{data.frame}

\item{by}{a collection of column names. See examples below}
}
\value{
dataframe which is joining y to x as of by columns
}
\description{
As of join
}
\examples{

op <- options(digits.secs = 6)
d1 <- data.frame(ts = c(Sys.time(),Sys.time()+1,Sys.time()+3), bid = c(100,99,100))
d2 <- data.frame(ts = c(Sys.time(),Sys.time()+2,Sys.time()+2), ask = c(101,100,102))
aj(ts,x=d1,y=d2)
}
