% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce.R
\name{coalesce}
\alias{coalesce}
\title{coalesce fills missing values}
\usage{
coalesce(table, value)
}
\arguments{
\item{table}{a dataframe}

\item{value}{a value}
}
\value{
dataframe whose missing values have been replaced by value
}
\description{
coalesce fills missing values
}
