% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross}
\alias{cross}
\title{cross is a cartesian product application of a function}
\usage{
cross(x, y, FUN = function(x, y) c(x, y), reduce = NULL)
}
\arguments{
\item{x}{a vector}

\item{y}{is a vector}

\item{FUN}{a binary function}
}
\value{
dataframe of application of FUN to cartesian product of x and y
}
\description{
cross is a cartesian product application of a function
}
