% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{dict}
\alias{dict}
\title{dict creates a named vector as a python like dictionary}
\usage{
dict(k, v)
}
\arguments{
\item{k}{vector of character}

\item{v}{vector of values}
}
\value{
named vector
}
\description{
dict creates a named vector as a python like dictionary
}
