% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{eval_char}
\alias{eval_char}
\title{Eval a character seens as an expression in the given environment}
\usage{
eval_char(s, env = parent.frame())
}
\arguments{
\item{s}{a char string}

\item{env}{an environment. Default is parent.frame()}
}
\value{

}
\description{
Eval a character seens as an expression in the given environment
}
\examples{
x<-2
eval_char("x^2 + 6")
}
