% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{Shortcut for select(where=..., by=...)}
\usage{
filter(d, ...)
}
\arguments{
\item{d}{data.frame}

\item{...}{a collection of boolean conditions on the columns}
}
\value{
the filtered data.frame
}
\description{
Shortcut for select(where=..., by=...)
}
\examples{
 d <- data.frame(
date=seq(as.Date("2001-01-01"), as.Date("2001-10-01"), by="month"),
ticker=sample(c("1 HK", "2 HK"), 10, TRUE),
ret1=rnorm(10),
ret2=rnorm(10),
ret3=rnorm(10)
)

d \%>\% filter(ret1 >0, ret2 <0)

}
