% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{first}
\alias{first}
\title{Pick first element}
\usage{
first(x)
}
\arguments{
\item{x}{any structure compatible with head}
}
\value{
first element
}
\description{
Pick first element
}
\examples{
x <- c(1,2,3)
first(x)
}
