% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{gather}
\alias{gather}
\title{Gather several columns into one}
\usage{
gather(d, ..., key = "key", value = "value")
}
\arguments{
\item{d}{data.frame}

\item{...}{colum names to gather}

\item{key}{a character name for the column containing the column names given in ...}

\item{value}{a character name for the column containing the values of the columns given in ...}
}
\value{
the transformed data.frame
}
\description{
Gather several columns into one
}
\examples{

d <- data.frame(
date=seq(as.Date("2001-01-01"), as.Date("2001-10-01"), by="month"),
ticker=sample(c("1 HK", "2 HK"), 10, TRUE),
ret1=rnorm(10),
ret2=rnorm(10),
ret3=rnorm(10)
)

d \%>\% gather(ret1,ret2)

d \%>\% gather(ret1,ret2,ret3, key="retType", value="retValue")

}
