% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{\%.\%}
\alias{\%.\%}
\title{Tensor product to construct parameter grids (similar to expand.grid)}
\usage{
x \%.\% y
}
\arguments{
\item{x}{a vector, list or data.frame}

\item{y}{a vector, list or data.frame}
}
\value{
a data.frame representing the tensor product of x \%.\% y
}
\description{
Tensor product to construct parameter grids (similar to expand.grid)
}
\examples{

1:3 \%.\% 1:3 #similar to expand.grid(1:3, 1:3)

1:3 \%.\% 1:3 \%.\% 1:4
}
