% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vect_pipe.R
\name{\%.>\%}
\alias{\%.>\%}
\title{Vectorized pipe}
\usage{
x \%.>\% f
}
\arguments{
\item{f}{}
}
\value{
a vector containing f(x_i) for all i
}
\description{
Vectorized pipe
}
\examples{
c(1,2,3) \%.>\% (\(x) x+1)
}
