% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{\%_\%}
\alias{\%_\%}
\title{remove first y values from x}
\usage{
x \%_\% y
}
\arguments{
\item{x}{a vector}

\item{y}{an integer}
}
\value{
vector with y values dropped from front of x
}
\description{
remove first y values from x
}
