% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject.R
\name{\%>>\%}
\alias{\%>>\%}
\title{Injection operator for dynamical creation of columns as parametrized images of a function (should be plugged in a select or an update)}
\usage{
x \%>>\% y
}
\arguments{
\item{x}{a list/vector or data.frame corresponding to the grid of parameters}

\item{y}{the parametrized function partially evaluated on columns under the form f(col1, col2,...)}
}
\value{
the alist containing the recipe to construct the new columns
}
\description{
Injection operator for dynamical creation of columns as parametrized images of a function (should be plugged in a select or an update)
}
\examples{
d <- data.frame(
date=seq(as.Date("2001-01-01"), as.Date("2001-10-01"), by="month"),
ticker=sample(c("1 HK", "2 HK"), 10, TRUE),
ret1=rnorm(10),
ret2=rnorm(10),
ret3=rnorm(10)
)

f <-function(x,y,k) (x*y)^k
d \%>\% select(1:5 \%>>\% f(ret1,ret2))

f <-function(x,y,k,l) (x*y)^(k+l)
d \%>\% select(1:2 \%.\% 1:3 \%>>\% f(ret1,ret2))
}
