% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep.R
\name{\%grep\%}
\alias{\%grep\%}
\title{grep function returning the matching values}
\usage{
pattern \%grep\% x
}
\arguments{
\item{pattern}{pattern to grep}

\item{x}{list/vector of strings}
}
\value{
the vector of matching values
}
\description{
grep function returning the matching values
}
\examples{
 "aa" \%grep\% c("a","aa", "aab", "caab")

}
