% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take.R
\name{\%#\%}
\alias{\%#\%}
\title{Infix shortcut for head}
\usage{
x \%#\% y
}
\arguments{
\item{x}{number of elements to pull out}

\item{y}{vector/list/data.frame...}
}
\value{
head(y,x)
}
\description{
Infix shortcut for head
}
\examples{
5 \%#\% 1:25
}
