% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{\%??\%}
\alias{\%??\%}
\title{Uniform sampling without replacement}
\usage{
x \%??\% y
}
\arguments{
\item{x}{sampling size}

\item{y}{vector/list}
}
\value{
subsampled  vector/list
}
\description{
Uniform sampling without replacement
}
\examples{
5 \%??\% c(1,3,4, 5 ,6,8)
}
