% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like.R
\name{\%like\%}
\alias{\%like\%}
\title{like operator similar to the SQL like keyword}
\usage{
x \%like\% y
}
\arguments{
\item{x}{a list/vector/data.frame}

\item{y}{a regular expression}
}
\value{
Boolean vector giving the position of matches
}
\description{
like operator similar to the SQL like keyword
}
