% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try.R
\name{\%try\%}
\alias{\%try\%}
\title{Try to compute an expression and returns default value if an error is raised}
\usage{
x \%try\% y
}
\arguments{
\item{x}{the default value}

\item{y}{the expression to evaluate}
}
\value{
the value to compute
}
\description{
Try to compute an expression and returns default value if an error is raised
}
\examples{

NA \%try\% (5+5)
}
