% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group}
\alias{group}
\title{split the dataframe into subgroups w.r.t some columns}
\usage{
group(d, ...)
}
\arguments{
\item{d}{data.frame}

\item{...}{columns}
}
\value{
list of data.frames
}
\description{
split the dataframe into subgroups w.r.t some columns
}
\examples{

d <- data.frame(
date=seq(as.Date("2001-01-01"), as.Date("2001-10-01"), by="month"),
ticker=sample(c("1 HK", "2 HK"), 10, TRUE),
ret1=rnorm(10),
ret2=rnorm(10),
ret3=rnorm(10)

d \%>\% group(ticker)
}
