% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{last}
\alias{last}
\title{Pick last element}
\usage{
last(x)
}
\arguments{
\item{x}{any structure compatible with head}
}
\value{
last element
}
\description{
Pick last element
}
\examples{
x <- c(1,2,3)
last(x)
}
