% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lj.R
\name{lj}
\alias{lj}
\title{SQL-like left join}
\usage{
lj(x, y, keep.y = FALSE)
}
\arguments{
\item{x}{left data.frame}

\item{y}{right data.frame}

\item{keep.y}{}
}
\value{
merged data.frame
}
\description{
SQL-like left join
}
