% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on.R
\name{on}
\alias{on}
\title{similar to the 'on' statement in SQL}
\usage{
on(d, ...)
}
\arguments{
\item{d}{data.frame}

\item{...}{column names}
}
\value{
a keyed data.frame (see xkey function)
}
\description{
similar to the 'on' statement in SQL
}
\examples{
d <- data.frame(
 date=seq(as.Date("2001-01-01"), as.Date("2001-10-01"), by="month"),
 ticker=sample(c("1 HK", "2 HK"), 10, TRUE),
 ret1=rnorm(10),
 ret2=rnorm(10),
 ret3=rnorm(10)
)

x <- data.frame(ticker = c("1 HK", "2 HK"), ndeals = c(100,500))
d \%>\% lj(x \%>\% on(ticker))
}
