% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parse_expr}
\alias{parse_expr}
\title{a shortcut for parse(text=...)[\link{1}]}
\usage{
parse_expr(t)
}
\arguments{
\item{t}{text to parse}
}
\value{
the parsed expression
}
\description{
a shortcut for parse(text=...)[\link{1}]
}
\examples{
x<-parse_expr("a=2")
eval(x)
}
