% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull}
\alias{pull}
\title{pulls ccolumns to the left of a datra.frame}
\usage{
pull(d, ...)
}
\arguments{
\item{d}{data frame}

\item{...}{columns}
}
\value{
the reorganized data frame
}
\description{
pulls ccolumns to the left of a datra.frame
}
\examples{

  d <- rbind(data.frame(date="20140101", ticker= 2, price=3),
  data.frame(date="20140102", ticker= 2, price=2),
  data.frame(date="20140101", ticker= 1, price=5))
  d \%>\% pull(ticker, price)
}
