% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{recycle}
\alias{recycle}
\title{Recycling value in given environment, for internal use}
\usage{
recycle(x, cols, env)
}
\arguments{
\item{env}{}
}
\value{

}
\description{
Recycling value in given environment, for internal use
}
