% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\title{Rename some columns of a data.frame}
\usage{
rename(d, ...)
}
\arguments{
\item{d}{data.frame}

\item{...}{'what' arguments: a collection of column names definition, e.g under the form new_name = old_name. See examples below}
}
\value{
renamed data.frame
}
\description{
Rename some columns of a data.frame
}
\examples{

d <- data.frame(
date=seq(as.Date("2001-01-01"), as.Date("2001-10-01"), by="month"),
ticker=sample(c("1 HK", "2 HK"), 10, TRUE),
ret1=rnorm(10),
ret2=rnorm(10),
ret3=rnorm(10)
)

d \%>\% rename(newret1 = ret1, newret2 = ret2)

It is also possible to pass an alist as follows

d \%>\% rename(newret1 = ret1, alist(newret2 = ret2, newret3 = ret3))

or a character list

d \%>\% rename(newret1 = ret1, list(newret2 = "ret2", newret3 = "ret3"))

}
