% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleby.R
\name{sampleby}
\alias{sampleby}
\title{similar to a select but by a given time frequency}
\usage{
sampleby(d, ..., where = NULL, by = NULL, freq = NULL, start = NULL)
}
\arguments{
\item{d}{a data.frame}

\item{...}{'what' arguments: a collection of column definitions. They often are aggregations of the original columns. See examples below}

\item{where}{boolean condition on columns, under the form cond1 if only one condition, and c(cond1, cond2, ...) if more than one condition}

\item{by}{the name of the 'timestamp' column in the data.frame from which time is sampled. Will raise an error if NULL or not one of d's column names}

\item{freq}{sampling frequency. Format is, e.g, 5m (5 minutes), 10s (10 seconds), 2d (two days), etc. See examples below}

\item{start}{start time in the frequency. If NULL (default value), start time is taken as the first timestamp encountered. Everything before the start time is dropped.}
}
\value{
the sampled data.frame
}
\description{
similar to a select but by a given time frequency
}
\examples{

Note: it is assumed that rows are ordered by the timestamp column provided in by.
Time format is any natural number followed by Y/M/d/h/m/s/mil/mic where mil and mic are respectively millisecond and microsecond.

d <- data.frame(ret1 = 1:10, ret2 = rnorm(10), time = Sys.time() + 60*c(1.1,2.2,3.3,4.4,4.5,6.6,7.7,8.8,9.9,10.1))

d \%>\% sampleby(ret1 = mean(ret1), ret2= mean(ret2), by = time, freq = "2m" )

d \%>\% sampleby(ret1 = mean(ret1), ret2= mean(ret2), by = time, freq = "2m", start = d[1,"time"] - 60)

}
