% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_aux}
\alias{select_aux}
\title{Auxiliary select function for internal use}
\usage{
select_aux(d, where = NULL, by = NULL, what = NULL, env = NULL)
}
\arguments{
\item{env}{}
}
\value{

}
\description{
Auxiliary select function for internal use
}
