% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread.R
\name{spread}
\alias{spread}
\title{Spread a column into several ones}
\usage{
spread(d, key, value)
}
\arguments{
\item{d}{data.frame}

\item{key}{the column containing labels of resulting columns}

\item{value}{the value column to spread}
}
\value{
the spread data.frame
}
\description{
Spread a column into several ones
}
\examples{
d <- data.frame(
date=seq(as.Date("2001-01-01"), as.Date("2001-10-01"), by="month"),
ticker=sample(c("1 HK", "2 HK"), 10, TRUE),
ret1=rnorm(10),
ret2=rnorm(10),
ret3=rnorm(10)
)

u <- d \%>\% gather(ret1,ret2)
u \%>\% spread(key,value)

}
