% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subvec}
\alias{subvec}
\title{negative index compatible slicing of a vector or a list}
\usage{
subvec(x, start, end)
}
\arguments{
\item{x}{anything compatible with []}

\item{start}{starting index (included). must be >= -length(x) and not 0}

\item{end}{end index (included). must be >= -length(x) and not 0}
}
\value{
the sliced vector. negative index n < 0 designates position length(x)+n+1. For instance, -1 designates length(x). index 0 is meaningless and will throw an error. If start and end have the same sign, the usual [] rule is applied. Otherwise, the vector is sliced by starting from start and 'walking up' to end assuming that the first element of x is the successor of the last element of x. See examples below.
}
\description{
negative index compatible slicing of a vector or a list
}
\examples{
x <- 1:10
x \%>\% subvec(-1,2) # c(10,1,2)
x \%>\% subvec(-7,-5) # c(4,5,6)
x \%>\% subvec(-5,-7) # c(6,5,4)
x \%>\% subvec(1,-7) # c(1,2,3,4)
x \%>\% subvec(1,-1) # 1:10
}
