% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{today}
\alias{today}
\title{today - n days in format YYYY-MM-DD}
\usage{
today(n = 0)
}
\arguments{
\item{n}{days shift from the current day. Defaults to 0.}
}
\value{
today - n days in format YYYY-MM-DD
}
\description{
today - n days in format YYYY-MM-DD
}
\examples{
today()
today(-1) # yesterday
today(1) # tomorrow
today(-7) # a week ago
}
