% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.R
\name{transpose}
\alias{transpose}
\title{Transpose a data.frame along some pivot variables}
\usage{
transpose(d, ...)
}
\arguments{
\item{d}{data.frame}

\item{...}{pivot variables whose values become columns. There the n-uplets thus formed should be unique and do not repeat in the data.}
}
\value{
the transposed data.frame
}
\description{
Transpose a data.frame along some pivot variables
}
\examples{
d <- data.frame(
date=seq(as.Date("2001-01-01"), as.Date("2001-10-01"), by="month"),
ticker=sample(c("1 HK", "2 HK"), 10, TRUE),
ret1=rnorm(10),
ret2=rnorm(10),
ret3=rnorm(10)
)

d[1,] \%>\% transpose
d \%>\% transpose(date)
d \%>\% transpose(date, ticker)

#these don't work:
d \%>\% transpose #pure transpose only works for one-row data.frame
d \%>\% transpose(ticker) #duplicated values in ticker

}
