% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{returns a boolean vector indicating which values to trim}
\usage{
trim(x, down = 0, up = 1)
}
\arguments{
\item{x}{vector o trim}

\item{down}{threshold down}

\item{up}{threshold up}
}
\value{

}
\description{
returns a boolean vector indicating which values to trim
}
\examples{


d <- data.frame(
date=seq(as.Date("2001-01-01"), as.Date("2001-10-01"), by="month"),
ticker=sample(c("1 HK", "2 HK"), 10, TRUE),
ret1=rnorm(10),
ret2=rnorm(10),
ret3=rnorm(10)
)

d \%>\% filter(ret1 \%>\% trim(0.1,0.9) )
}
