% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ungroup.R
\name{ungroup}
\alias{ungroup}
\title{Merge a list of data.frame having the same keys and columns}
\usage{
ungroup(l)
}
\arguments{
\item{l}{list of data.frame to merge}
}
\value{
merged data.frame
}
\description{
Merge a list of data.frame having the same keys and columns
}
