% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{unparse}
\alias{unparse}
\title{Deparse x, but avoid introducing artificial breaks in the returned character string}
\usage{
unparse(x)
}
\arguments{
\item{x}{value to deparse}
}
\value{
deparse(x) with no breaks
}
\description{
Deparse x, but avoid introducing artificial breaks in the returned character string
}
