% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{SQL-like update function with internal "where" and "by" features}
\usage{
update(d, ..., where = NULL, by = NULL)
}
\arguments{
\item{d}{data.frame}

\item{...}{'what' arguments: a collection of column definitions. See examples below}

\item{where}{boolean condition on columns, under the form cond1 if only one condition, and c(cond1, cond2, ...) if more than one condition}

\item{by}{collection of columns, under the form col1 if only one condition, and c(col1, col2, ...) if more than one column}
}
\value{
updated data.frame
}
\description{
SQL-like update function with internal "where" and "by" features
}
\examples{

d <- data.frame(
date=seq(as.Date("2001-01-01"), as.Date("2001-10-01"), by="month"),
ticker=sample(c("1 HK", "2 HK"), 10, TRUE),
ret1=rnorm(10),
ret2=rnorm(10),
ret3=rnorm(10)
)

d \%>\% update(vol = mean(ret1^2))

d \%>\% update(vol1 = mean(ret1^2), vol2 = mean(ret2^2), where = c(date!="2001-05-01", ret3 >0), by =ticker)

}
