% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_aux}
\alias{update_aux}
\title{Auxiliary update function, for internal use}
\usage{
update_aux(d, where = NULL, by = NULL, what = NULL, env = NULL)
}
\arguments{
\item{env}{}
}
\value{

}
\description{
Auxiliary update function, for internal use
}
