% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmean.R
\name{wmean}
\alias{wmean}
\title{quick weighted mean with na.rm=T}
\usage{
wmean(x, w = NULL, ...)
}
\arguments{
\item{x}{vector}

\item{w}{weight vector}

\item{...}{other params passed to mean}
}
\value{

}
\description{
quick weighted mean with na.rm=T
}
\examples{

 x<- c(NA,2,3,4)
 w <- c(1,NA,2,2)
 wmean(x,w)
}
