% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsd.R
\name{wsd}
\alias{wsd}
\title{quick weighted sd with na.rm=T}
\usage{
wsd(x, w = NULL, ...)
}
\arguments{
\item{x}{vector}

\item{w}{weight vector}

\item{...}{other params passed to sd/mean}
}
\value{

}
\description{
quick weighted sd with na.rm=T
}
\examples{

 x<- c(NA,2,3,4)
 w <- c(1,NA,2,2)
 wsd(x,w)
}
