% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xasc.R
\name{xasc}
\alias{xasc}
\title{SQL-like ascending order. Sort is by the first column given, then by the second column within the first, and so on.}
\usage{
xasc(d, cols)
}
\arguments{
\item{d}{data.frame}

\item{cols}{char vector of column names}
}
\value{
sorted data.frame
}
\description{
SQL-like ascending order. Sort is by the first column given, then by the second column within the first, and so on.
}
\examples{
d <- data.frame(
date=seq(as.Date("2001-01-01"), as.Date("2001-10-01"), by="month"),
ticker=sample(c("1 HK", "2 HK"), 10, TRUE),
ret1=rnorm(10),
ret2=rnorm(10),
ret3=rnorm(10)
)

d \%>\% xasc("ret1")
}
