% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcol.R
\name{xcol}
\alias{xcol}
\title{Rename columns of a data.frame}
\usage{
xcol(d, cols)
}
\arguments{
\item{d}{data.frame}

\item{cols}{vector of new column names}
}
\value{
renamed data.frame
}
\description{
Rename columns of a data.frame
}
\examples{

d <- data.frame(
date=seq(as.Date("2001-01-01"), as.Date("2001-10-01"), by="month"),
ticker=sample(c("1 HK", "2 HK"), 10, TRUE),
ret1=rnorm(10),
ret2=rnorm(10),
ret3=rnorm(10)
)

d \%>\% xcol(c("date" , "return1"))
}
