% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgroup.R
\name{xgroup}
\alias{xgroup}
\title{Split a data.frame into a list of data.frame by values of a subset of columns}
\usage{
xgroup(d, cols = NULL)
}
\arguments{
\item{d}{data.frame}

\item{cols}{character vector of columns with respect to which perform the group operation.}
}
\value{
a list of sub-blocks data.frames following the grouping rule
}
\description{
Split a data.frame into a list of data.frame by values of a subset of columns
}
\examples{

d <- data.frame(
date=seq(as.Date("2001-01-01"), as.Date("2001-10-01"), by="month"),
ticker=sample(c("1 HK", "2 HK"), 10, TRUE),
ret1=rnorm(10),
ret2=rnorm(10),
ret3=rnorm(10)
)

d \%>\% xgroup("ticker")
}
