% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xprev.R
\name{xprev}
\alias{xprev}
\title{Return the vector with an offset of k indices}
\usage{
xprev(l, k)
}
\arguments{
\item{l}{list/vector}

\item{k}{offset}
}
\value{
offset list/vector
}
\description{
Return the vector with an offset of k indices
}
\examples{

x <- c(1,3,5)
xprev(x,1)
xprev(x,-1)
}
